refactor/16-move-anniversary-reminder-to-hr_addon_settings.py-v15
